+++ c++/src/vm/assembler/assembler.hpp
@@ -26,7 +26,7 @@ class VmAssembler
     VmAssembler(std::string code, bool debug = false);
     void assamble();
     void dumpToBinary(std::string name);
-    std::vector<uint64_t> loadFromBinary(std::string name);
+    std::vector<uint8_t> loadFromBinary(std::string name);
     
     // Expose symbols and instructions for linker
     std::vector<Instruction> getInstructions() const;
